pyproject.toml
[project]
name = "lil-gnida"
version = "0.1.0"
description = "A telegram chat-bot for using different local-hosted stuff"
authors = [
    {name = "neerayel",email = "andrei11shum@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot (>=22.5,<23.0)",
    "pydantic (>=2.12.3,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "aiohttp (>=3.13.1,<4.0.0)"
]

[tool.poetry]
packages = [{include = "lil_gnida", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "mypy (>=1.18.2,<2.0.0)"
]
